Semana16Hackaton/fmillaa/requirements.txt
django
gunicorn
django-heroku
djangorestframework
django-filter
markdown
coverage
pytest-django